refactor: LottoWithBonus to vo package
