+++ src/test/java/lotto/domain/vo/LottoWithBonusTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.List;
 import lotto.domain.constants.ErrorCode;
+import lotto.domain.vo.Lotto;
 import org.junit.jupiter.api.Test;
 
 class LottoWithBonusTest {
